pyproject.toml
[project]
name = "Neurodivergente-Backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Fabrica-da-resenha", email = "fabrica.resenha.idmfabio@gmail.com"},
]
dependencies = ["django>=5.2.7", "djangorestframework>=3.16.1", "django-cors-headers>=4.9.0", "gunicorn>=23.0.0", "djangorestframework-simplejwt>=5.5.1", "pillow>=11.3.0"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}

[tool.poetry]
name = "neurodivergente-backend"
version = "0.1.0"
description = ""
authors = ["Fabrica-da-Resenha <fabrica.resenha.idmfabio@gmail.com>"]
readme = "README.md"
packages = [
    { include = "config", from = "src" },
    { include = "core", from = "src" },
]

[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "python src/manage.py runserver"
createsuperuser = "python src/manage.py createsuperuser"
pre_migrate = "python src/manage.py makemigrations"
migrate = "python src/manage.py migrate"
shell = "python src/manage.py shell"
test = "python src/manage.py test"